internal: Add modified `container/ring` pkg
We're going to implement the "clock" as a linked list.
The standard library provides `container/ring` but we can modify it to
take advantage of generic type parameters and iterators that have been
added to Go after `container/ring`'s API was defined.

In addition, we can store LIRS metadata directly within the `Ring`,
rather than require our implementation to explicitly wrap values.